+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # Unreleased
 
-## Added
-
 ## Fixed
 
-## Changed
+- Catch an exception in the notifier plugin which can occur in headless (CI)
+  setups
 
 # 1.62.993 (2022-02-04 / 083f69e)